{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2022.7.1  # https://github.com/stub42/pytz
python-slugify==8.0.0  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
# BUG: prevents Django Compressor from installing on windows.  I was able to install rcssmin==1.1.1  without issues
# {#%- if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %#}
# rcssmin==1.1.0 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
# {%- else %}
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
# {#%- endif %#}
{%- endif %}
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==6.3.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==4.4.2  # https://github.com/redis/redis-py
{%- if cookiecutter.use_docker == "y" or cookiecutter.windows == "n" %}
hiredis==2.1.1  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.4.0  # https://github.com/celery/django-celery-beat
{%- if cookiecutter.use_docker == 'y' %}
flower==1.2.0  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if cookiecutter.use_async == 'y' %}
uvicorn[standard]==0.20.0  # https://github.com/encode/uvicorn
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
djangorestframework-guardian # https://github.com/rpkilby/django-rest-framework-guardian
django-guardian # https://github.com/django-guardian/django-guardian
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# {%- if cookiecutter.use_drf == 'y' %}
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0  # https://github.com/adamchainz/django-cors-headers
drf-jwt
PyJWT
# DRF-spectacular for api documentation
drf-spectacular==0.25.1  # https://github.com/tfranzel/drf-spectacular
{%- endif %}
{%- if cookiecutter.frontend_pipeline == 'Webpack' %}
django-webpack-loader==1.8.0  # https://github.com/django-webpack/django-webpack-loader
{%- endif %}

djangoql
django-json-widget
django-picklefield
{%- if cookiecutter.cloud_provider == 'AWS' %}
boto3
{%- endif %}


